+++ src/components/inputField/InputField.styled.ts
@@ -1,0 +1,33 @@
+import styled from 'styled-components';
+
+export const Container = styled.div`
+  margin-top: 10px;
+`;
+
+export const Title = styled.div`
+  font-size: 24px;
+  font-weight: 700;
+  margin: 0;
+`;
+
+export const Subtitle = styled.div`
+  font-size: 13px;
+  font-weight: 300;
+  margin: 10px 0 12px 0;
+  color: #8b95a1;
+`;
+
+export const Label = styled.label`
+  font-size: 15.5px;
+  font-weight: 400;
+  display: block;
+  margin-top: 24px;
+  color: #0a0d13;
+`;
+
+export const InputWrapper = styled.div`
+  display: flex;
+  width: 440px;
+  justify-content: space-between;
+  margin-top: 8px;
+`;
